mcp-kusto-azure-function-example/requirements.txt
azure-functions
azure-kusto-data
mcp
